+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.1.0] - 2026-07-10
+
+- Rebranded from AuSync to ClauSync
+- Updated bundle ID to com.darkinlife71.clausync
+- Updated all references and branding
+
 ## [1.0.0] - 2026-07-08
 
 - Rebranded Telegram Drive as AuSync